requirements.txt
cloudscraper==1.2.60
requests==2.31.0
prettytable==3.6.0
colorama==0.4.6
beautifulsoup4==4.12.2
urllib3==2.0.7
PySocks==1.7.1